+++ latch_cli/services/cp/main.py
@@ -1,5 +1,8 @@
 from pathlib import Path
+from textwrap import dedent
 from typing import List
+
+import click
 
 from latch_cli.services.cp.config import CPConfig, Progress
 from latch_cli.services.cp.download import download
@@ -41,8 +44,12 @@ def cp(
             else:
                 remote_copy(src, dest)
         else:
-            raise ValueError(
-                f"`latch cp` cannot be used for purely local file copying. Please make"
-                f" sure one or both of your paths is a remote path (beginning with"
-                f" `latch://`)"
+            click.secho(
+                dedent(f"""
+                `latch cp` cannot be used for purely local file copying.
+
+                Please ensure at least one of your arguments is a remote path (beginning with `latch://`)
+                """).strip("\n"),
+                fg="red",
             )
+            raise click.exceptions.Exit(1)