+++ latch_cli/services/move.py
@@ -30,7 +30,7 @@ def move(
         srcs = expand_pattern(src)
 
     if len(srcs) == 0:
-        click.secho(f"Could not find any files that match pattern {src}. Exiting.")
+        click.echo(f"Could not find any files that match pattern {src}. Exiting.")
         raise click.exceptions.Exit(0)
 
     node_data = get_node_data(*srcs, dest, allow_resolve_to_parent=True)
@@ -55,19 +55,21 @@ def move(
         )
         raise click.exceptions.Exit(1)
 
+    path_by_id = {v.id: k for k, v in node_data.data.items()}
+
     for s in srcs:
         src_data = node_data.data[s]
 
-        path_by_id = {v.id: k for k, v in node_data.data.items()}
-
         if src_data.is_parent:
-            raise get_path_error(s, "not found", acc_id)
+            click.echo(get_path_error(s, "not found", acc_id))
+            raise click.exceptions.Exit(1)
 
         new_name = None
         if dest_data.is_parent:
             new_name = get_name_from_path(dest)
         elif dest_data.type in {LDataNodeType.obj, LDataNodeType.link}:
-            raise get_path_error(dest, "object already exists at path.", acc_id)
+            click.echo(get_path_error(dest, "object already exists at path.", acc_id))
+            raise click.exceptions.Exit(1)
 
         try:
             execute(
@@ -104,33 +106,44 @@ def move(
                 node_id = msg.rsplit(" ", 1)[1]
                 path = path_by_id[node_id]
 
-                raise get_path_error(path, "permission denied.", acc_id) from e
+                click.echo(get_path_error(path, "permission denied.", acc_id))
+                raise click.exceptions.Exit(1) from e
             elif msg == "Refusing to make node its own parent":
-                raise get_path_error(dest, f"is a parent of {s}.", acc_id) from e
+                click.echo(get_path_error(dest, f"is a parent of {s}.", acc_id))
+                raise click.exceptions.Exit(1) from e
             elif msg == "Refusing to parent node to an object node":
-                raise get_path_error(dest, f"object exists at path.", acc_id) from e
+                click.echo(get_path_error(dest, f"object exists at path.", acc_id))
+                raise click.exceptions.Exit(1) from e
             elif msg == "Refusing to move a share link (or into a share link)":
                 if src_data.type is LDataNodeType.link:
                     path = s
                 else:
                     path = dest
 
-                raise get_path_error(path, f"is a share link.", acc_id) from e
+                click.echo(get_path_error(path, f"is a share link.", acc_id))
+                raise click.exceptions.Exit(1) from e
             elif msg.startswith("Refusing to move account root"):
-                raise get_path_error(s, "is an account root.", acc_id) from e
+                click.echo(get_path_error(s, "is an account root.", acc_id))
+                raise click.exceptions.Exit(1) from e
             elif msg.startswith("Refusing to move removed node"):
-                raise get_path_error(s, "not found.", acc_id) from e
+                click.echo(get_path_error(s, "not found.", acc_id))
+                raise click.exceptions.Exit(1) from e
             elif msg.startswith("Refusing to move already moved node"):
-                raise get_path_error(
-                    s,
-                    "copy in progress. Please wait until the node has finished copying"
-                    " before moving.",
-                    acc_id,
-                ) from e
+                click.echo(
+                    get_path_error(
+                        s,
+                        "copy in progress. Please wait until the node has finished"
+                        " copying before moving.",
+                        acc_id,
+                    )
+                )
+                raise click.exceptions.Exit(1) from e
             elif msg == "Conflicting object in destination":
-                raise get_path_error(dest, "object exists at path.", acc_id) from e
+                click.echo(get_path_error(dest, "object exists at path.", acc_id))
+                raise click.exceptions.Exit(1) from e
             elif msg.startswith("Refusing to do noop move"):
-                raise get_path_error(dest, "cannot move node to itself.", acc_id) from e
+                click.echo(get_path_error(dest, "cannot move node to itself.", acc_id))
+                raise click.exceptions.Exit(1) from e
             else:
                 raise e